# controllers/bookController.go
package controllers

import (
	"assignment1/models"
	"database/sql"
	"encoding/json"
	"fmt"
	"log"
	"net/http"
	"strconv"

	"github.com/gorilla/mux"
)

// @Summary Get all books
// @Description Retrieves a list of all books.
// @Produce json
// @Success 200 {array} models.Book
// @Router /books [get]
func GetBooks(db *sql.DB) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		rows, err := db.Query("SELECT id, title, author, year FROM books")
		if err != nil {
			http.Error(w, err.Error(), http.StatusInternalServerError)
			return
		}
		defer rows.Close()
		var books []models.Book
		for rows.Next() {
			var book models.Book
			err := rows.Scan(&book.ID, &book.Title, &book.Author, &book.Year)
			if err != nil {
				http.Error(w, err.Error(), http.StatusInternalServerError)
				return
			}
			books = append(books, book)
		}
		w.WriteHeader(http.StatusOK)
		json.NewEncoder(w).Encode(books)
	}
}

// @Summary Get a book by ID
// @Description Retrieves a book based on provided ID.
// @Produce json
// @Param id path string true "Book ID"
// @Success 200 {object} models.Book
// @Router /books/{id} [get]
func GetBookById(db *sql.DB) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		vars := mux.Vars(r)
		log.Println(vars)
		id, err := strconv.Atoi(vars["id"])
		if err != nil {
			http.Error(w, err.Error(), http.StatusBadRequest)
			return
		}
		fmt.Println(id)
		var book models.Book
		row := db.QueryRow(`SELECT id, title, author, year FROM books WHERE id=$1`, id)
		log.Println(row, "row")
		err = row.Scan(&book.ID, &book.Title, &book.Author, &book.Year)
		if err == sql.ErrNoRows {
			http.Error(w, "Book not found", http.StatusNotFound)
			return
		} else if err != nil {
			http.Error(w, err.Error(), http.StatusInternalServerError)
			return
		}
		log.Println("book response", book)
		w.WriteHeader(http.StatusOK)
		json.NewEncoder(w).Encode(book)
	}
}

// @Summary Create a new book
// @Description Creates a new book with provided data.
// @Accept json
// @Produce json
// @Param book body models.Book true "Book object"
// @Success 201 {object} models.Book
// @Router /books [post]
func CreateBook(db *sql.DB) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		var book models.Book
		err := json.NewDecoder(r.Body).Decode(&book)
		if err != nil {
			http.Error(w, err.Error(), http.StatusBadRequest)
			return
		}
		var id int
		err = db.QueryRow("INSERT INTO books (title, author, year) VALUES ($1, $2, $3) RETURNING id", book.Title, book.Author, book.Year).Scan(&id)
		if err != nil {
			http.Error(w, err.Error(), http.StatusInternalServerError)
			return
		}
		log.Println("id", id)
		book.ID = int(id)
		w.WriteHeader(http.StatusCreated)
		json.NewEncoder(w).Encode(map[string]interface{}{
			"message": "Book created successfully",
			"book":    book,
		})
	}
}

// @Summary Update an existing book
// @Description Updates an existing book with provided data.
// @Accept json
// @Produce json
// @Param id path string true "Book ID"
// @Param book body models.Book true "Book object"
// @Success 200 {object} models.Book
// @Router /books/{id} [put]
func UpdateBook(db *sql.DB) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		vars := mux.Vars(r)
		id, err := strconv.Atoi(vars["id"])
		if err != nil {
			http.Error(w, "Invalid book ID", http.StatusBadRequest)
			return
		}

		var count int
		err = db.QueryRow("SELECT COUNT(*) FROM books WHERE id = $1", id).Scan(&count)
		if err != nil {
			http.Error(w, err.Error(), http.StatusInternalServerError)
			return
		}
		if count == 0 {
			http.Error(w, "No record found", http.StatusNotFound)
			return
		}
		log.Println("count", count)
		var book models.Book
		err = json.NewDecoder(r.Body).Decode(&book)
		if err != nil {
			http.Error(w, err.Error(), http.StatusBadRequest)
			return
		}

		_, err = db.Exec("UPDATE books SET title=$1, author=$2, year=$3 WHERE id=$4", book.Title, book.Author, book.Year, id)
		if err != nil {
			http.Error(w, err.Error(), http.StatusInternalServerError)
			return
		}
		book.ID = id
		w.WriteHeader(http.StatusOK)
		json.NewEncoder(w).Encode(map[string]interface{}{
			"message": "Book updated successfully",
			"book":    book,
		})
	}
}

// @Summary Delete a book
// @Description Deletes a book based on provided ID.
// @Param id path string true "Book ID"
// @Success 204 "No Content"
// @Router /books/{id} [delete]
func DeleteBook(db *sql.DB) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		vars := mux.Vars(r)
		id, err := strconv.Atoi(vars["id"])
		if err != nil {
			http.Error(w, "Invalid book ID", http.StatusBadRequest)
			return
		}

		var count int
		err = db.QueryRow("SELECT COUNT(*) FROM books WHERE id = $1", id).Scan(&count)
		if err != nil {
			http.Error(w, err.Error(), http.StatusInternalServerError)
			return
		}
		if count == 0 {
			http.Error(w, "No record found", http.StatusNotFound)
			return
		}

		_, err = db.Exec("DELETE FROM books WHERE id=$1", id)
		if err != nil {
			http.Error(w, err.Error(), http.StatusInternalServerError)
			return
		}

		w.WriteHeader(http.StatusOK)
		json.NewEncoder(w).Encode(map[string]interface{}{
			"message": "Book deleted successfully",
			"book_id": id,
		})
	}
}